refactor: move vehicle related use cases to a package
