+++ src/domain/use_cases/vehicle/publish_vehicle.py
